hive_data_extraction_service/requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
PyHive
sasl==0.2.1
thrift==0.10.0
thrift-sasl==0.3.0
pylibmc
cython
requests
